[add] cancel auto clean unused image and set to API
